Remove 100B win bonus cap; migrate wins to NUMERIC for arbitrary precision
The streak bonus was capped at 10B (100B with max bonus_mult) to prevent
BIGINT overflow. wins column is now NUMERIC (arbitrary precision in Postgres)
so the cap is no longer needed.

Changes:
- schema.sql: wins BIGINT → NUMERIC
- game.py: remove min(..., 10_000_000_000) cap from bonus calculation
- game.py: wrap gs['wins'] / r['wins'] with int() in all three endpoints
  that read it (state, spin, leaderboard) — psycopg2 returns Decimal for
  NUMERIC columns; int() converts to Python arbitrary-precision int which
  JSON-serialises correctly
- Live DB: ALTER TABLE game_state ALTER COLUMN wins TYPE NUMERIC (applied)

Python's int is already arbitrary precision, so no server-side overflow
is possible. Postgres NUMERIC has no practical upper bound.

+++ schema.sql
@@ -16,7 +16,7 @@ CREATE UNIQUE INDEX IF NOT EXISTS idx_users_device ON users(device_id) WHERE dev
 
 CREATE TABLE IF NOT EXISTS game_state (
     user_id        INTEGER PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
-    wins           BIGINT NOT NULL DEFAULT 0,
+    wins           NUMERIC NOT NULL DEFAULT 0,
     losses         BIGINT NOT NULL DEFAULT 0,
     fish_clicks    BIGINT NOT NULL DEFAULT 0,
     streak         INTEGER NOT NULL DEFAULT 0,
